A variety of improvements for groups

+++ app/assets/javascripts/groups.js.coffee
@@ -101,13 +101,14 @@ window.GroupSettingsView = class GroupSettingsView
     @group_id = options.group_id
     @el = $('#groups-settings')
     @el.find('.remove-member').click (e) ->
-      $.ajax
-        url: "/memberships/#{$(e.currentTarget).attr('data-membership-id')}.json"
-        type: "DELETE"
-        data: 
-          authenticity_token : $('meta[name=csrf-token]').attr('content')              
-        success: (x) =>
-          $(e.currentTarget).parents('tr').remove()
+      if confirm("Knoda is even more fun with friends. Do you really remove this user from the group?")
+        $.ajax
+          url: "/memberships/#{$(e.currentTarget).attr('data-membership-id')}.json"
+          type: "DELETE"
+          data: 
+            authenticity_token : $('meta[name=csrf-token]').attr('content')              
+          success: (x) =>
+            $(e.currentTarget).parents('tr').remove()
     $('body')
       .on 'focus', '[contenteditable]', ->
           $this = $(this)
